+++ scripts/git-subtree/chinese-convert/add.ps1
@@ -1,0 +1,5 @@
+git subtree add `
+    -P workspace/plugins/chinese-convert `
+    https://github.com/Zuoqiu-Yingyi/siyuan-plugin-chinese-convert.git `
+    dev `
+    --squash