+++ scripts/git-subtree/chinese-convert/pull.ps1
@@ -1,0 +1,4 @@
+git subtree pull `
+    -P workspace/plugins/chinese-convert `
+    https://github.com/Zuoqiu-Yingyi/siyuan-plugin-chinese-convert.git `
+    dev