+++ scripts/git-subtree/chinese-convert/push.ps1
@@ -1,0 +1,4 @@
+git subtree push `
+    -P workspace/plugins/chinese-convert `
+    https://github.com/Zuoqiu-Yingyi/siyuan-plugin-chinese-convert.git `
+    dev